Move NearbyPage to dataclient package
