+++ app/src/main/java/org/wikipedia/places/PlacesFragmentViewModel.kt
@@ -1,6 +1,5 @@
 package org.wikipedia.places
 
-import android.graphics.Bitmap
 import android.location.Location
 import android.os.Bundle
 import androidx.lifecycle.MutableLiveData
@@ -9,10 +8,10 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.launch
-import org.maplibre.android.plugins.annotation.Symbol
 import org.wikipedia.Constants
 import org.wikipedia.dataclient.ServiceFactory
 import org.wikipedia.dataclient.WikiSite
+import org.wikipedia.dataclient.page.NearbyPage
 import org.wikipedia.extensions.parcelable
 import org.wikipedia.page.PageTitle
 import org.wikipedia.settings.Prefs
@@ -46,23 +45,6 @@ class PlacesFragmentViewModel(bundle: Bundle) : ViewModel() {
                     }
                 }
             nearbyPagesLiveData.postValue(Resource.Success(pages))
-        }
-    }
-
-    class NearbyPage(
-        val pageId: Int,
-        val pageTitle: PageTitle,
-        val latitude: Double,
-        val longitude: Double,
-        var annotation: Symbol? = null,
-        var bitmap: Bitmap? = null
-    ) {
-
-        private val lat = latitude
-        private val lng = longitude
-        val location get() = Location("").apply {
-            latitude = lat
-            longitude = lng
         }
     }
